Refactor project structure and update references
Refactored project structure to move DTOs and interfaces to the Application layer. Updated project references and namespaces accordingly. Modified tests and services to use the new locations of DTOs and interfaces. Updated directory tree to reflect the new structure.

Clean Arch compliance - round 2.
