+++ TeamWeeklyStatus.Application/DTOs/DoneThisWeekTaskDTO.cs
@@ -4,6 +4,7 @@
     {
         public int Id { get; set; }
         public string TaskDescription { get; set; }
+        public List<SubtaskDTO> Subtasks { get; set; } = new List<SubtaskDTO>();
     }
 
 }
