+++ TeamWeeklyStatus.Application/DTOs/TeamMemberDTO.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace TeamWeeklyStatus.Domain.DTOs
+namespace TeamWeeklyStatus.Application.DTOs
 {
     public class TeamMemberDTO
     {
