+++ TeamWeeklyStatus.Application/Services/WeeklyStatusService.cs
@@ -1,7 +1,6 @@
-﻿using TeamWeeklyStatus.Application.Interfaces;
-using TeamWeeklyStatus.Domain.DTOs;
+﻿using TeamWeeklyStatus.Application.DTOs;
+using TeamWeeklyStatus.Application.Interfaces;
 using TeamWeeklyStatus.Domain.Entities;
-using TeamWeeklyStatus.Infrastructure.Repositories;
 
 namespace TeamWeeklyStatus.Application.Services
 {
