--- TeamWeeklyStatus.Domain/DTOs/DoneThisWeekTaskDTO.cs
@@ -1,8 +1,0 @@
-﻿namespace TeamWeeklyStatus.Domain.DTOs
-{
-    public class DoneThisWeekTaskDTO
-    {
-        public string TaskDescription { get; set; }
-        public List<SubtaskDTO> Subtasks { get; set; } = new List<SubtaskDTO>();
-    }
-}
