--- TeamWeeklyStatus.Domain/DTOs/TaskWithSubtasksDTO.cs
@@ -1,8 +1,0 @@
-﻿namespace TeamWeeklyStatus.Domain.DTOs
-{
-    public class TaskWithSubtasksDTO
-    {
-        public string TaskDescription { get; set; }
-        public List<string> Subtasks { get; set; } = new List<string>();
-    }
-}
