+++ TeamWeeklyStatus.Infrastructure/Repositories/Repository.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using TeamWeeklyStatus.Application.Interfaces;
 
 namespace TeamWeeklyStatus.Infrastructure.Repositories
 {
