+++ TeamWeeklyStatus.WebApi/Controllers/WeeklyStatusController.cs
@@ -41,14 +41,14 @@ namespace TeamWeeklyStatus.WebApi.Controllers
         [HttpPost("Add", Name = "SaveWeeklyStatus")]
         public async Task<IActionResult> SaveWeeklyStatus([FromBody] WeeklyStatusPostRequest request)
         {
-            var weeklyStatusDto = new Domain.DTOs.WeeklyStatusDTO
+            var weeklyStatusDto = new Application.DTOs.WeeklyStatusDTO
             {
                 MemberId = request.MemberId,
                 WeekStartDate = request.WeekStartDate,
-                DoneThisWeek = request.DoneThisWeek.Select(dtw => new Domain.DTOs.DoneThisWeekTaskDTO
+                DoneThisWeek = request.DoneThisWeek.Select(dtw => new Application.DTOs.DoneThisWeekTaskDTO
                 {
                     TaskDescription = dtw.TaskDescription,
-                    Subtasks = dtw.Subtasks.Select(sub => new Domain.DTOs.SubtaskDTO { SubtaskDescription = sub.SubtaskDescription }).ToList()
+                    Subtasks = dtw.Subtasks.Select(sub => new Application.DTOs.SubtaskDTO { SubtaskDescription = sub.SubtaskDescription }).ToList()
                 }).ToList(),
                 PlanForNextWeek = request.PlanForNextWeek,
                 Blockers = request.Blockers,
@@ -62,15 +62,15 @@ namespace TeamWeeklyStatus.WebApi.Controllers
         [HttpPut("Edit", Name = "UpdateWeeklyStatus")]
         public async Task<IActionResult> UpdateWeeklyStatus([FromBody] WeeklyStatusPostRequest request)
         {
-            var weeklyStatusDto = new Domain.DTOs.WeeklyStatusDTO
+            var weeklyStatusDto = new Application.DTOs.WeeklyStatusDTO
             {
                 Id = request.Id,
                 MemberId = request.MemberId,
                 WeekStartDate = request.WeekStartDate,
-                DoneThisWeek = request.DoneThisWeek.Select(dtw => new Domain.DTOs.DoneThisWeekTaskDTO
+                DoneThisWeek = request.DoneThisWeek.Select(dtw => new Application.DTOs.DoneThisWeekTaskDTO
                 {
                     TaskDescription = dtw.TaskDescription,
-                    Subtasks = dtw.Subtasks.Select(sub => new Domain.DTOs.SubtaskDTO { SubtaskDescription = sub.SubtaskDescription }).ToList()
+                    Subtasks = dtw.Subtasks.Select(sub => new Application.DTOs.SubtaskDTO { SubtaskDescription = sub.SubtaskDescription }).ToList()
                 }).ToList(),
                 PlanForNextWeek = request.PlanForNextWeek,
                 Blockers = request.Blockers,
